refactor: Extract into files.
