+++ src/action.rs
@@ -7,13 +7,6 @@ use bevy::ecs::{
 };
 
 /// Send this event.
-pub fn send_event<E: Event + Clone>(event: E) -> impl FnMut(EventWriter<E>) {
-    move |mut writer: EventWriter<E>| {
-        writer.send(event.clone());
-    }
-}
-
-/// Send an event if a condition is met.
 ///
 /// ```
 /// use bevy::prelude::*;
@@ -25,26 +18,12 @@ pub fn send_event<E: Event + Clone>(event: E) -> impl FnMut(EventWriter<E>) {
 /// struct MyEvent;
 ///
 /// KeySequence::new(
-///    action::send_event_if(MyEvent, in_state(AppState::Game)),
+///    action::send_event(MyEvent),
 ///    keyseq! { Space });
 /// ```
-pub fn send_event_if<E: Event + Clone, M>(
-    event: E,
-    condition: impl Condition<M>,
-) -> impl FnMut(&mut World) {
-    let mut condition_system = Some(IntoSystem::into_system(condition));
-    let mut system_id: Option<SystemId<(), bool>> = None;
-    move |world: &mut World| {
-        if system_id.is_none() {
-            system_id =
-                Some(world.register_system(condition_system.take().expect("No condition system")));
-        }
-        if world
-            .run_system(system_id.expect("Condition not registered"))
-            .expect("Condition run failed")
-        {
-            world.send_event(event.clone());
-        }
+pub fn send_event<E: Event + Clone>(event: E) -> impl FnMut(EventWriter<E>) {
+    move |mut writer: EventWriter<E>| {
+        writer.send(event.clone());
     }
 }
