+++ src/lib.rs
@@ -29,340 +29,22 @@ pub use keyseq::{
     Modifiers,
 };
 
+mod cache;
+mod plugin;
 pub mod action;
 pub mod cond_system;
 mod covec;
 mod frame_time;
 mod input_sequence;
 mod time_limit;
+mod chord;
+
+pub use plugin::{InputSequencePlugin};
+pub use cache::{InputSequenceCache};
+pub use chord::KeyChord;
 
 use covec::Covec;
 use frame_time::FrameTime;
-
-/// Represents a key chord, i.e., a set of modifiers and a key code.
-#[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Reflect)]
-pub struct KeyChord(pub Modifiers, pub KeyCode);
-
-impl fmt::Display for KeyChord {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        use KeyCode::*;
-        self.0.fmt(f)?;
-        if !self.0.is_empty() {
-            f.write_str("-")?;
-        }
-        let key_repr = match self.1 {
-            Semicolon => ";",
-            Period => ".",
-            Equal => "=",
-            Slash => "/",
-            Minus => "-",
-            BracketLeft => "[",
-            BracketRight => "]",
-            Quote => "'",
-            Backquote => "`",
-            key_code => {
-                let mut key = key_code.variant_name();
-                key = key.strip_prefix("Key").unwrap_or(key);
-                key = key.strip_prefix("Digit").unwrap_or(key);
-                return f.write_str(key);
-            }
-        };
-        f.write_str(key_repr)
-    }
-}
-
-impl From<(Modifiers, KeyCode)> for KeyChord {
-    #[inline(always)]
-    fn from((mods, key): (Modifiers, KeyCode)) -> Self {
-        KeyChord(mods, key)
-    }
-}
-
-impl From<KeyCode> for KeyChord {
-    #[inline(always)]
-    fn from(key: KeyCode) -> Self {
-        KeyChord(Modifiers::empty(), key)
-    }
-}
-
-/// Input sequence plugin.
-pub struct InputSequencePlugin {
-    #[allow(clippy::type_complexity)]
-    schedules: Vec<(Interned<dyn ScheduleLabel>, Option<Interned<dyn SystemSet>>)>,
-}
-
-impl Default for InputSequencePlugin {
-    fn default() -> Self {
-        InputSequencePlugin {
-            schedules: vec![(Interned(Box::leak(Box::new(Update))), None)],
-        }
-    }
-}
-
-impl Plugin for InputSequencePlugin {
-    fn build(&self, app: &mut App) {
-        if app.world.get_resource::<Input<KeyCode>>().is_some() {
-            // Add key sequence.
-            app.init_resource::<InputSequenceCache<KeyChord, ()>>();
-
-            for (schedule, set) in &self.schedules {
-                if let Some(set) = set {
-                    app.add_systems(
-                        *schedule,
-                        (
-                            detect_removals::<KeyChord, ()>,
-                            detect_additions::<KeyChord, ()>,
-                            key_sequence_matcher,
-                        )
-                            .chain()
-                            .in_set(*set),
-                    );
-                } else {
-                    app.add_systems(
-                        *schedule,
-                        (
-                            detect_removals::<KeyChord, ()>,
-                            detect_additions::<KeyChord, ()>,
-                            key_sequence_matcher,
-                        )
-                            .chain(),
-                    );
-                }
-            }
-        } else {
-            warn!("No key sequence matcher added; consider adding DefaultPlugins.");
-        }
-
-        if app.world.get_resource::<Input<GamepadButton>>().is_some() {
-            // Add button sequences.
-            app.init_resource::<InputSequenceCache<GamepadButtonType, Gamepad>>();
-
-            for (schedule, set) in &self.schedules {
-                if let Some(set) = set {
-                    app.add_systems(
-                        *schedule,
-                        (
-                            detect_removals::<GamepadButtonType, Gamepad>,
-                            detect_additions::<GamepadButtonType, Gamepad>,
-                            button_sequence_matcher,
-                        )
-                            .chain()
-                            .in_set(*set),
-                    );
-                } else {
-                    app.add_systems(
-                        *schedule,
-                        (
-                            detect_removals::<GamepadButtonType, Gamepad>,
-                            detect_additions::<GamepadButtonType, Gamepad>,
-                            button_sequence_matcher,
-                        )
-                            .chain(),
-                    );
-                }
-            }
-        } else {
-            warn!("No button sequence matcher added; consider adding DefaultPlugins.");
-        }
-    }
-}
-
-impl InputSequencePlugin {
-    /// Constructs an empty input sequence plugin with no default schedules.
-    pub fn empty() -> Self {
-        Self { schedules: vec![] }
-    }
-    /// Run the executor in a specific `Schedule`.
-    pub fn run_in(mut self, schedule: impl ScheduleLabel) -> Self {
-        self.schedules
-            .push((Interned(Box::leak(Box::new(schedule))), None));
-        self
-    }
-
-    /// Run the executor in a specific `Schedule` and `SystemSet`.
-    pub fn run_in_set(mut self, schedule: impl ScheduleLabel, set: impl SystemSet) -> Self {
-        self.schedules.push((
-            Interned(Box::leak(Box::new(schedule))),
-            Some(Interned(Box::leak(Box::new(set)))),
-        ));
-        self
-    }
-}
-
-/// Contains the trie for the input sequences.
-#[derive(Resource)]
-pub struct InputSequenceCache<A, In> {
-    trie: Option<Trie<A, InputSequence<A, In>>>,
-}
-
-impl<A, In> InputSequenceCache<A, In>
-where
-    A: Ord + Clone + Send + Sync + 'static,
-    In: Send + Sync + Clone + 'static,
-{
-    /// Retrieve the cached trie without iterating through `sequences`. Or if
-    /// the cache has been invalidated, build and cache a new trie using the
-    /// `sequences` iterator.
-    pub fn trie<'a>(
-        &mut self,
-        sequences: impl Iterator<Item = &'a InputSequence<A, In>>,
-    ) -> &Trie<A, InputSequence<A, In>> {
-        self.trie.get_or_insert_with(|| {
-            let mut builder: TrieBuilder<A, InputSequence<A, In>> = TrieBuilder::new();
-            for sequence in sequences {
-                builder.insert(sequence.acts.clone(), sequence.clone());
-            }
-            builder.build()
-        })
-    }
-}
-
-impl<A, In> Default for InputSequenceCache<A, In> {
-    fn default() -> Self {
-        Self { trie: None }
-    }
-}
-
-/// The systems run in one of these sets.
-#[derive(SystemSet, Debug, Clone, PartialEq, Eq, Hash)]
-pub enum InputSet {
-    /// [KeySequences] run in this set.
-    Key,
-    /// [ButtonSequences] run in this set.
-    Button,
-}
-
-fn is_modifier(key: KeyCode) -> bool {
-    let mods = Modifiers::from(key);
-    !mods.is_empty()
-}
-
-#[allow(clippy::too_many_arguments)]
-fn button_sequence_matcher(
-    secrets: Query<&ButtonSequence>,
-    time: Res<Time>,
-    buttons: Res<Input<GamepadButton>>,
-    mut last_buttons: Local<HashMap<usize, Covec<GamepadButtonType, FrameTime>>>,
-    mut cache: ResMut<InputSequenceCache<GamepadButtonType, Gamepad>>,
-    frame_count: Res<FrameCount>,
-    mut commands: Commands,
-) {
-    let trie = cache.trie(secrets.iter());
-    let now = FrameTime {
-        frame: frame_count.0,
-        time: time.elapsed_seconds(),
-    };
-    for button in buttons.get_just_pressed() {
-        let pad_buttons = match last_buttons.get_mut(&button.gamepad.id) {
-            Some(x) => x,
-            None => {
-                last_buttons.insert(button.gamepad.id, Covec::default());
-                last_buttons.get_mut(&button.gamepad.id).unwrap()
-            }
-        };
-
-        pad_buttons.push(button.button_type, now.clone());
-        let start = pad_buttons.1[0].clone();
-        for seq in consume_input(trie, &mut pad_buttons.0) {
-            if seq
-                .time_limit
-                .as_ref()
-                .map(|limit| (&now - &start).has_timedout(limit))
-                .unwrap_or(false)
-            {
-                // Sequence timed out.
-            } else {
-                commands.run_system_with_input(seq.system_id, button.gamepad);
-            }
-        }
-        pad_buttons.drain1_sync();
-    }
-}
-
-#[allow(clippy::too_many_arguments)]
-fn key_sequence_matcher(
-    secrets: Query<&KeySequence>,
-    time: Res<Time>,
-    keys: Res<Input<KeyCode>>,
-    mut last_keys: Local<Covec<KeyChord, FrameTime>>,
-    mut cache: ResMut<InputSequenceCache<KeyChord, ()>>,
-    frame_count: Res<FrameCount>,
-    mut commands: Commands,
-) {
-    let mods = Modifiers::from_input(&keys);
-    let trie = cache.trie(secrets.iter());
-    let now = FrameTime {
-        frame: frame_count.0,
-        time: time.elapsed_seconds(),
-    };
-    for key_code in keys.get_just_pressed() {
-        if is_modifier(*key_code) {
-            continue;
-        }
-        let key = KeyChord(mods, *key_code);
-        last_keys.push(key, now.clone());
-        let start = last_keys.1[0].clone();
-        for seq in consume_input(trie, &mut last_keys.0) {
-            if seq
-                .time_limit
-                .as_ref()
-                .map(|limit| (&now - &start).has_timedout(limit))
-                .unwrap_or(false)
-            {
-                // Sequence timed out.
-            } else {
-                commands.run_system(seq.system_id);
-            }
-        }
-        last_keys.drain1_sync();
-    }
-}
-
-fn detect_additions<A: Clone + Send + Sync + 'static, In: 'static>(
-    secrets: Query<&InputSequence<A, In>, Added<InputSequence<A, In>>>,
-    mut cache: ResMut<InputSequenceCache<A, In>>,
-) {
-    if secrets.iter().next().is_some() {
-        cache.trie = None;
-    }
-}
-
-fn detect_removals<A: Clone + Send + Sync + 'static, In: 'static>(
-    mut cache: ResMut<InputSequenceCache<A, In>>,
-    mut removals: RemovedComponents<InputSequence<A, In>>,
-) {
-    if removals.read().next().is_some() {
-        cache.trie = None;
-    }
-}
-
-fn consume_input<'a, K, V>(trie: &'a Trie<K, V>, input: &mut Vec<K>) -> impl Iterator<Item = &'a V>
-where
-    K: Clone + Eq + Ord,
-    // V: Clone,
-{
-    let mut result = vec![];
-    let mut min_prefix = None;
-    for i in 0..input.len() {
-        if let Some(seq) = trie.exact_match(&input[i..]) {
-            result.push(seq);
-        }
-        if min_prefix.is_none() && trie.is_prefix(&input[i..]) {
-            min_prefix = Some(i);
-            // let _ = input.drain(0..i);
-            // return result.into_iter();
-        }
-    }
-    match min_prefix {
-        Some(i) => {
-            let _ = input.drain(0..i);
-        }
-        None => {
-            input.clear();
-        }
-    }
-    result.into_iter()
-}
 
 #[cfg(test)]
 mod tests {
